Dev 1

+++ accreapp/views.py
@@ -76,6 +76,7 @@ ORDER_COLUMN_CHOICES =Choices(
     ('4', 'tagged_items__tag__code'),
     ('5', 'uploaded_at'),
     ('6', 'id'),
+    ('7', 'uploaded_by')
 )
 
 def mainTableData(request):
@@ -97,15 +98,16 @@ def mainTableData(request):
         prof = Profile.objects.get(pk= user.id)
         prof_department =  prof.department
 
-        allfiles = File.objects.all().distinct()
-        if request.user.is_staff and not request.user.is_superuser:
-            allfiles = File.objects.filter(user__profile__department = prof_department).all().distinct()
+        allfiles = File.objects.all().exclude(flag=1).distinct()
+        #if request.user.is_staff and not request.user.is_superuser:
+        #    allfiles = File.objects.filter(user__profile__department = prof_department).all().distinct()
         total = allfiles.count()
         
         if search_value:
             allfiles = allfiles.filter(Q(file_name__icontains=search_value) |
                                     Q(tags__code__icontains=search_value) |
-                                    Q(description__icontains=search_value)).all().distinct()
+                                    Q(user__username__icontains=search_value) |
+                                    Q(description__icontains=search_value)).all().exclude(flag=1).distinct()
         count = allfiles.count()
         if length != -1:
             allfiles = allfiles.order_by(order_column)[start:start + length]
@@ -123,6 +125,7 @@ def mainTableData(request):
                 'tagged_items__tag__code': i.get_tags(),
                 'date_created': dd,
                 'description': i.description,
+                'uploaded_by': i.user.username,
                 'uploaded_at': i.uploaded_at
             })
 
@@ -149,7 +152,8 @@ def bulkUpdate(request):
         is_updated = 1
         d = {}
         content_type = ContentType.objects.get(id=9)
-        
+        your_ids = []
+
         if len(controlNums) > 0 :
             if not description == '':
                 d['description'] = description
@@ -158,15 +162,21 @@ def bulkUpdate(request):
             bulk_taggedFile = []
             for i in controlNums:
                 b = File.objects.get(id=i)
+                if b.user.id == current_user.id:
+                    your_ids.append(i)
                 codes = []
                 for j in tags:
                     if len(tags) > 0 and tags[0] != '':
-                        if Category.objects.filter(id=j).exists():
-                            tagss =  Category.objects.get(id = j)
+                        tagss =  Category.objects.get(id = j)
+                        if Category.objects.filter(id=j).exists() and b.user.id == current_user.id:
                             TaggedWhatever.objects.filter(object_id=i).delete()
                             bulk_taggedFile.append(TaggedWhatever(object_id = i,content_type = content_type,tag = tagss))
                             TaggedWhatever.objects.bulk_create(bulk_taggedFile)
                             codes.append(tagss.code)      
+                        elif Category.objects.filter(id=j).exists():
+                            bulk_taggedFile.append(TaggedWhatever(object_id = i,content_type = content_type,tag = tagss))
+                            TaggedWhatever.objects.bulk_create(bulk_taggedFile)
+                            codes.append(tagss.code) 
                         else:
                             print ('invalid tag')
                 if len(tags) > 0 and tags[0] != '':
@@ -178,7 +188,8 @@ def bulkUpdate(request):
                     action.send(user_instance, verb='Updated',action_object=b,description='Updated File  '+str(b.file_name)+'\'s description.' , target=content_type)
                 elif not date_created == '':     
                     action.send(user_instance, verb='Updated',action_object=b,description='Updated File  '+str(b.file_name)+'\'s date.' , target=content_type)
-            File.objects.filter(id__in=controlNums).update(**d)
+            print (your_ids)
+            File.objects.filter(id__in=your_ids).update(**d)
            
         return JsonResponse({'is_updated': is_updated})
     else:
@@ -202,9 +213,27 @@ def bulkDelete(request):
                 action.send(user_instance, verb='Deleted',action_object=b,description='Deleted File  '+str(b.file_name) , target=content_type)
             if counter == len(controlNums):
                 print ('wwwww')
-                File.objects.filter(id__in=controlNums).delete()
+                File.objects.filter(id__in=controlNums).update(flag=1)
             return JsonResponse({'is_deleted_items': is_deleted})
         else:
             return JsonResponse({'is_deleted_items': 0})
+    else:
+        raise Http404
+
+
+def removeTag(request):
+    current_user = request.user
+    user_instance = User.objects.get(id=current_user.id)
+    content_type = ContentType.objects.get(id=13)
+
+    if request.user.is_authenticated:
+        tag = request.POST.get('tag')
+        file = request.POST.get('file_name')
+        b = File.objects.get(file_name=file)
+        print (tag)
+        TaggedWhatever.objects.filter(tag__code=tag).delete()
+        action.send(user_instance, verb='Removed tag',action_object=b,description='Removed tag  '+tag+' on '+str(b.file_name) , target=content_type)
+
+        return JsonResponse({'is_deleted_items': 1})
     else:
         raise Http404
